+++ ricaun.Revit.Templates/content/ProjectTemplates/ricaun.Project.AutoCAD.Addin/AutoCAD/Commands/Command.cs
@@ -1,0 +1,15 @@
+using Autodesk.AutoCAD.ApplicationServices;
+using Autodesk.AutoCAD.Runtime;
+using System;
+
+namespace ProjectTemplates.AutoCAD.Commands
+{
+    public static class Command
+    {
+        [CommandMethod("ThemeChange")]
+        public static void ThemeChange()
+        {
+            Application.SetSystemVariable("COLORTHEME", 1 ^ (short)Application.GetSystemVariable("COLORTHEME"));
+        }
+    }
+}